+++ app/src/main/java/com/demo/desh/model/RetrofitDto.kt
@@ -1,5 +1,6 @@
 package com.demo.desh.model
 
+import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
 data class RecommendInfo(
@@ -19,3 +20,19 @@ data class ServiceList(
     val size: Int,
     val list: List<String>
 ): Serializable
+
+data class User(
+    var id : Long? = null,
+    @SerializedName("nickname") val nickname : String,
+    @SerializedName("email") val email : String,
+    @SerializedName("profileImageUrl") val profileImageUrl : String,
+) : Serializable
+
+data class Realty(
+    @SerializedName("name") val name: String,
+    @SerializedName("price") val price: Double,
+    @SerializedName("address") val address: String,
+    @SerializedName("pyung") val pyung: Long,
+    @SerializedName("squareMeter") val squareMeter: Double,
+    @SerializedName("userId") val userId: Long
+) : Serializable
